+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.1.0] - 2024-XX-XX
+New release of the HyperDbg Debugger.
+
+### Added
+- Added feature to pause the debuggee immediately upon connection
+
+### Changed
+
 ## [0.10.0.0] - 2024-07-22
 New release of the HyperDbg Debugger.